Sway 0.45 (#26)
* remaded all into sway 0.45

* moved to fuels 0.46

* integrated token factory

* removed token abi

* import src20_sdk from github

---------

+++ contracts/market/tests/utils/number_utils.rs
@@ -1,8 +1,8 @@
 // TODO: Add the ability to accept floats
-pub fn parse_units(num: u64, decimals: u8) -> u64 {
+pub fn parse_units(num: u64, decimals: u64) -> u64 {
     num * 10u64.pow(decimals as u32)
 }
 
-pub fn format_units(num: u64, decimals: u8) -> f64 {
+pub fn format_units(num: u64, decimals: u64) -> f64 {
     num as f64 / 10u64.pow(decimals as u32) as f64
 }
